requirements.txt
google-adk
google-generativeai
PyYAML
python-dotenv
requests
